+++ net/minecraft/world/level/levelgen/structure/templatesystem/AxisAlignedLinearPosTest.java
@@ -42,7 +42,7 @@ extends PosRuleTest {
         float h = Math.abs((blockPos2.getZ() - blockPos3.getZ()) * direction.getStepZ());
         int i = (int)(f + g + h);
         float j = random.nextFloat();
-        return (double)j <= Mth.clampedLerp(this.minChance, this.maxChance, Mth.inverseLerp(i, this.minDist, this.maxDist));
+        return (double)j <= Mth.clampedLerp((double)this.minChance, (double)this.maxChance, Mth.inverseLerp(i, this.minDist, this.maxDist));
     }
 
     @Override
